fix vmap text
